#278: Split rendering of each fixture-date into own component
