+++ CourageScores/ClientApp/src/components/division_fixtures/DivisionFixtures.js
@@ -3,7 +3,6 @@ import {DivisionFixture} from "./DivisionFixture";
 import {NewFixtureDate} from "./NewFixtureDate";
 import {Dialog} from "../common/Dialog";
 import {ProposeGamesDialog} from "./ProposeGamesDialog";
-import {TournamentFixture} from "./TournamentFixture";
 import {NewTournamentGame} from "./NewTournamentGame";
 import {FilterFixtures} from "./FilterFixtures";
 import {useLocation, useNavigate} from "react-router-dom";
@@ -12,8 +11,7 @@ import {any, isEmpty, stateChanged} from "../../Utilities";
 import {useDependencies} from "../../IocContainer";
 import {useApp} from "../../AppContainer";
 import {useDivisionData} from "../DivisionDataContainer";
-import { getFilters, isInPast, isToday } from "./FilterUtilities";
-import {FixtureDateNote} from "./FixtureDateNote";
+import {DivisionFixtureDate} from "./DivisionFixtureDate";
 
 export function DivisionFixtures({ setNewFixtures }) {
     const { id: divisionId, season, fixtures, teams, onReloadDivision } = useDivisionData();
@@ -271,75 +269,19 @@ export function DivisionFixtures({ setNewFixtures }) {
         setNewDate('');
     }
 
-    function hasProposals(fixtures) {
-        return any(fixtures, f => f.proposal);
-    }
-
-    function toggleShowPlayers(date) {
-        const newShowPlayers = Object.assign({}, showPlayers);
-        if (newShowPlayers[date]) {
-            delete newShowPlayers[date];
-        } else {
-            newShowPlayers[date] = true;
-        }
-        setShowPlayers(newShowPlayers);
-
-        navigate({
-            pathname: location.pathname,
-            search: location.search,
-            hash: any(Object.keys(newShowPlayers))
-                ? 'show-who-is-playing'
-                : '',
-        });
-    }
-
     function renderFixtureDate(date) {
-        const filters = getFilters(filter, renderContext, fixtures);
-        let fixturesForDate = (date.fixtures || []).filter(f => filters.apply({ date: date.date, fixture: f, tournamentFixture: false }));
-        const tournamentFixturesForDate = (date.tournamentFixtures || []).filter(f => filters.apply({ date: date.date, fixture: f, tournamentFixture: true }));
-        const notesForDate = date.notes;
-
-        const hasFixtures = any(date.fixtures, f => f.id !== f.homeTeam.id);
-        if (!isAdmin && !hasFixtures) {
-            fixturesForDate = []; // no fixtures defined for this date, and not an admin so none can be defined, hide all the teams
-        }
-
-        if (isEmpty(fixturesForDate) && isEmpty(tournamentFixturesForDate) && isEmpty(notesForDate)) {
-            return null;
-        }
-
-        return (<div key={date.date} className={isToday(date.date) ? 'text-primary' : (isInPast(date.date) || hasProposals(date.fixtures) ? '' : 'text-secondary-50')}>
-            <h4>
-                📅 {new Date(date.date).toDateString()}{date.hasKnockoutFixture ? (<span> (Qualifier)</span>) : null}
-                {isNoteAdmin ? (<button className="btn btn-primary btn-sm margin-left" onClick={() => startAddNote(date.date)}>📌 Add note</button>) : null}
-                {any(tournamentFixturesForDate) ? (
-                    <span className="margin-left form-switch h6 text-body">
-                        <input type="checkbox" className="form-check-input align-baseline"
-                               id={'showPlayers_' + date.date} checked={showPlayers[date.date] || false} onChange={() => toggleShowPlayers(date.date)} />
-                        <label className="form-check-label margin-left" htmlFor={'showPlayers_' + date.date}>Who's playing?</label>
-                    </span>) : null}
-            </h4>
-            {notesForDate.map(note => (<FixtureDateNote note={note} setEditNote={setEditNote} />))}
-            <table className="table layout-fixed">
-                <tbody>
-                {fixturesForDate.map(f => (<DivisionFixture
-                    key={f.id}
-                    fixture={f}
-                    readOnly={proposingGames}
-                    date={date.date}
-                    allowTeamDelete={false}
-                    allowTeamEdit={false}
-                    isKnockout={f.isKnockout}
-                    onUpdateFixtures={(apply) => setNewFixtures(apply(fixtures))} />))}
-                {tournamentFixturesForDate.map(tournament => (<TournamentFixture
-                    key={tournament.address + '-' + tournament.date}
-                    tournament={tournament}
-                    date={date.date}
-                    onTournamentChanged={onTournamentChanged}
-                    expanded={showPlayers[date.date]} />))}
-                </tbody>
-            </table>
-        </div>);
+        return (<DivisionFixtureDate
+            date={date}
+            filter={filter}
+            renderContext={renderContext}
+            proposingGames={proposingGames}
+            showPlayers={showPlayers}
+            startAddNote={startAddNote}
+            setEditNote={setEditNote}
+            setShowPlayers={setShowPlayers}
+            setNewFixtures={setNewFixtures}
+            onTournamentChanged={onTournamentChanged}
+        />);
     }
 
     function startAddNote(date) {
